+++ apps/dashboard/src/lib/components/Org/LandingPage/Testimonial.svelte
@@ -1,0 +1,12 @@
+<script>
+  import TestimonialCard from './components/TestimonialCard.svelte';
+</script>
+
+<section class="px-4 pt-4 pb-20 h-full bg-white">
+  <h1 class="text-center text-3xl">Student testimonial</h1>
+  <section class="flex gap-8 p-4">
+    <TestimonialCard />
+    <TestimonialCard />
+    <TestimonialCard />
+  </section>
+</section>
